ps3netsrv 20200527A + webMAN MOD 1.47.27
- Display PSP icon on /mount.ps3 screen when mounted from /net
- Increased some 8-bit length variables to 16bit in VisoFile.c of ps3netsrv
- Improved performace of scrncpy_upper (copy & check in a single loop)

+++ _Projects_/ps3netsrv/include/VIsoFile.h
@@ -14,8 +14,8 @@
 typedef struct _FileList
 {
 	char *path;
-	uint8_t path_len;
-	uint8_t file_len;
+	uint16_t path_len;
+	uint16_t file_len;
 	uint32_t rlba;
 	off64_t size;
 	bool multipart;
@@ -25,8 +25,8 @@ typedef struct _FileList
 typedef struct _DirList
 {
 	char *path;
-	uint8_t path_len;
-	uint8_t full_len;
+	uint16_t path_len;
+	uint16_t full_len;
 	uint8_t *content;
 	uint8_t *contentJoliet;
 	size_t contentSize;
